move nav helpers into separate file
